enhance migration

+++ database/migrations/2021_04_29_055701_create_wallets_table.php
@@ -21,10 +21,10 @@ class CreateWalletsTable extends Migration
             $table->foreign('auction_id')->references('id')->on('auctions')->onDelete('cascade');
             $table->unsignedBigInteger('currency_id');
             $table->foreign('currency_id')->references('id')->on('currencies')->onDelete('cascade');
-            $table->double('in');
-            $table->double('out');
-            $table->double('hold');
-            $table->double('balance');
+            $table->decimal('in',18);
+            $table->decimal('out',18);
+            $table->decimal('hold',18);
+            $table->decimal('balance',18);
 
             $table->timestamps();
         });
